backend/requirements.txt
# backend/requirements.txt — Python Dependencies
# ---------------------------------------------------------------------------
# Install with:  pip install -r requirements.txt
# Requires Python 3.11+
# ---------------------------------------------------------------------------

fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0
google-cloud-vision>=3.7.0
python-dotenv>=1.0.0
